Rename package to FormBuilder
