+++ app/webpacker/components/CompetitionForm/FormSections/Admin.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import { InputBoolean } from '../../wca/FormProvider/input/FormInputs';
+import { InputBoolean } from '../../wca/FormBuilder/input/FormInputs';
 import { useStore } from '../../../lib/providers/StoreProvider';
-import SubSection from '../../wca/FormProvider/SubSection';
+import SubSection from '../../wca/FormBuilder/SubSection';
 
 export default function Admin() {
   const { isAdminView, isPersisted } = useStore();
