+++ app/webpacker/components/CompetitionForm/FormSections/EventRestrictions.js
@@ -6,11 +6,11 @@ import {
   InputNumber,
   InputSelect,
   InputTextArea,
-} from '../../wca/FormProvider/input/FormInputs';
+} from '../../wca/FormBuilder/input/FormInputs';
 import { useStore } from '../../../lib/providers/StoreProvider';
 import ConditionalSection from './ConditionalSection';
-import SubSection from '../../wca/FormProvider/SubSection';
-import { useFormObject } from '../../wca/FormProvider/provider/FormObjectProvider';
+import SubSection from '../../wca/FormBuilder/SubSection';
+import { useFormObject } from '../../wca/FormBuilder/provider/FormObjectProvider';
 
 export default function EventRestrictions() {
   const {
