+++ app/webpacker/components/CompetitionForm/FormSections/NameDetails.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import { InputString } from '../../wca/FormProvider/input/FormInputs';
+import { InputString } from '../../wca/FormBuilder/input/FormInputs';
 import { useStore } from '../../../lib/providers/StoreProvider';
 import { competitionMaxShortNameLength } from '../../../lib/wca-data.js.erb';
-import { useFormObject } from '../../wca/FormProvider/provider/FormObjectProvider';
+import { useFormObject } from '../../wca/FormBuilder/provider/FormObjectProvider';
 
 export default function NameDetails() {
   const { isPersisted, isAdminView } = useStore();
