+++ app/webpacker/components/CompetitionForm/FormSections/VenueInfo.js
@@ -10,8 +10,8 @@ import {
   InputNumber,
   InputSelect,
   InputString,
-} from '../../wca/FormProvider/input/FormInputs';
-import SubSection from '../../wca/FormProvider/SubSection';
+} from '../../wca/FormBuilder/input/FormInputs';
+import SubSection from '../../wca/FormBuilder/SubSection';
 
 const countriesOptions = Object.values(countries.byIso2).map((country) => ({
   key: country.id,
