+++ app/webpacker/components/CompetitionForm/Tables/CompsTable.js
@@ -7,7 +7,7 @@ import I18n from '../../../lib/i18n';
 import { events } from '../../../lib/wca-data.js.erb';
 import useToggleState from '../../../lib/hooks/useToggleState';
 import { useStore } from '../../../lib/providers/StoreProvider';
-import { useFormObject } from '../../wca/FormProvider/provider/FormObjectProvider';
+import { useFormObject } from '../../wca/FormBuilder/provider/FormObjectProvider';
 
 function NotConfirmedIcon() {
   return (
